Remove bot reactions on arbitrary messages
Moderation actions will be confirmed by the bot reacting to its own
messages and a moderator also reacting to the message.  For security, this
requires removing the ability for arbitrary messages to receive bot
reactions, as that could enable a bot message to be reacted to without
explicitly sending the message via a command.