commands: pdf_image: add more aggressive resource limits
The memory instance of the bot is fairly limited. This change
applies much more aggressive limits on the page ranges to avoid
the wrath of the OOM killer. It also only decodes one page
at a time, to avoid storing large amounts in memory.
The resulting implementation is uglier, but keeps
the bot more responsive as well.

Internal representation has been switched to JPEG for more compression